updates more insistent, with lockfile. draft folder links correct, so no edit button needed. UI tweaks to settings.dtl purge the queue once SMP succeeds

+++ thirtythirty/management/commands/update.py
@@ -3,10 +3,13 @@ from django.core.management.base import BaseCommand
 
 from optparse import make_option
 
+import os.path
+
 import addressbook
 import emailclient
 
-import thirtythirty.updater as TTU
+import thirtythirty.utils as TTUtil
+import thirtythirty.updater as TTUp
 import thirtythirty.settings as TTS
 
 import logging
@@ -41,12 +44,13 @@ class Command(BaseCommand):
     def alert_user_new_updates(self,
                                Version=None,
                                ChangeLog=None):
+        if os.path.exists(TTS.UPSTREAM['update_lock']): return
         Me = addressbook.utils.my_address()
         PL = """
         A newer version of LookingGlass is available!
         
         To install it, go to "Settings" in the navbar (next to LOCKDOWN).
-        Open the "System administration" tab.
+        Open the "Updates" tab.
         Click "Update".
         ~CROSS YOUR FINGERS~
         ~WAIT FOR POSSIBLE REBOOT~
@@ -64,34 +68,32 @@ class Command(BaseCommand):
             Destination=Me.email,
             Subject='New update available: %s' % Version,
             From='Sysop <root>')
+        TTUtil.popen_wrapper(['/bin/touch',
+                               TTS.UPSTREAM['update_lock']],
+                              sudo=False)
 
-    def handle(self, *args, **settings):
-        Cached = TTU.Available()
-        if Cached:
-            if not settings['force']:
-                logger.debug('Already have version %s in cache' % Cached['version'])
-                exit(0)
-            else:
-                logger.debug('Already have version %s in cache' % Cached['version'])
-                logger.debug('Specified new dl anyway')
-            
-        URI = TTU.Scan(TTS.UPSTREAM['updates'][0]['uri'])
+    def handle(self, *args, **settings):            
+        URI = TTUp.Scan(TTS.UPSTREAM['updates'][0]['uri'])
         if not URI and not settings['force']:
             logger.debug('Already up to date')
             exit(0)
 
-        Cache = TTU.Update_Cache(Data_URI=URI,
-                                 Checksum_URI='%s.sum.asc' % URI)
+        if not settings['force']:
+            # we probably have something genuinely email-worthy to talk about
+            os.unlink(TTS.UPSTREAM['update_lock'])
+
+        Cache = TTUp.Update_Cache(Data_URI=URI,
+                                  Checksum_URI='%s.sum.asc' % URI)
         if Cache:
             logger.debug('Got %s' % Cache)
 
-        Cache = TTU.Available()
+        Cache = TTUp.Available()
         if not Cache:
             logger.warning('Nothing new in cache - bailing')
             exit()
         logger.debug('Reloading from cache: %s' % Cache['version'])
         
-        if not TTU.Validate(Cache['filename']):
+        if not TTUp.Validate(Cache['filename']):
             if not settings['valid_override']:
                 logger.warning('Invalid signature on %s' % Cache['filename'])
                 exit(-1)
@@ -99,12 +101,12 @@ class Command(BaseCommand):
                 logger.warning('Bad signature, but signature override is on...')
 
         self.alert_user_new_updates(
-            Version=TTU.Version(Cache['filename']),
-            ChangeLog=TTU.ChangeLog(Cache['filename'])
+            Version=TTUp.Version(Cache['filename']),
+            ChangeLog=TTUp.ChangeLog(Cache['filename'])
             )
 
         if settings['install']:
-            for F in TTU.Unpack(Cache['filename']):
+            for F in TTUp.Unpack(Cache['filename']):
                 logger.debug(F)
                 
-            TTU.Cleanup()
+            TTUp.Cleanup()
